refactor: reorganize entrypoints and libs
