+++ src/lib/utils/normalizeUrl.ts
